pyproject.toml
[project]
name = "spinal-curvature-estimation"
version = "0.1.0"
description = "A project during a 2-month research internship at ETH Zurich"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "addict==2.4.0",
    "albucore==0.0.24",
    "albumentations==2.0.8",
    "annotated-types==0.7.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "ccimport==0.4.4",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "configargparse==1.7.1",
    "contourpy==1.3.2",
    "cumm-cu118==0.7.11",
    "cumm-cu121==0.7.11",
    "cycler==0.12.1",
    "dash==3.1.1",
    "fastjsonschema==2.21.1",
    "filelock==3.18.0",
    "fire==0.7.0",
    "flask==3.1.1",
    "fonttools==4.58.5",
    "fsspec==2025.5.1",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "imageio==2.37.0",
    "importlib-metadata==8.7.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "jsonschema==4.24.1",
    "jsonschema-specifications==2025.4.1",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.8",
    "lark==1.2.2",
    "lazy-loader==0.4",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mpmath==1.3.0",
    "narwhals==1.47.1",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "ninja==1.11.1.4",
    "numpy==2.2.6",
    "open3d==0.19.0",
    "opencv-python-headless==4.11.0.86",
    "packaging==25.0",
    "pandas==2.3.1",
    "pccm==0.4.16",
    "pillow==11.3.0",
    "pip>=25.1.1",
    "platformdirs==4.3.8",
    "plotly==6.2.0",
    "portalocker==3.2.0",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "pybind11==3.0.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "pyquaternion==0.9.9",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.4",
    "retrying==1.4.0",
    "rpds-py==0.26.0",
    "safetensors==0.6.2",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "seaborn==0.12.2",
    "sentry-sdk==2.32.0",
    "simsimd==6.4.10",
    "six==1.17.0",
    "smmap==5.0.2",
    "spconv-cu118==2.3.8",
    "spconv-cu121==2.3.8",
    "stringzilla==3.12.5",
    "sympy==1.13.1",
    "termcolor==3.1.0",
    "threadpoolctl==3.6.0",
    "tifffile==2025.5.10",
    "timm==1.0.19",
    "torch==2.5.1+cu121",
    "torch-scatter==2.1.2+pt25cu121",
    "torchvision==0.20.1+cu121",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wandb==0.21.0",
    "werkzeug==3.1.3",
    "zipp==3.23.0",
]
